#396: Introduce OperationFactory.
